@@ -1,7 +1,7 @@
 ﻿using SQLite;
 using System.Collections;
 
-namespace v2rayN.Base
+namespace v2rayN
 {
     public sealed class SqliteHelper
     {
@@ -11,10 +11,11 @@ namespace v2rayN.Base
         private SQLiteConnection _db;
         private SQLiteAsyncConnection _dbAsync;
         private static readonly object objLock = new();
+        public readonly string _configDB = "guiNDB.db";
 
         public SqliteHelper()
         {
-            _connstr = Utils.GetConfigPath(Global.ConfigDB);
+            _connstr = Utils.GetConfigPath(_configDB);
             _db = new SQLiteConnection(_connstr, false);
             _dbAsync = new SQLiteAsyncConnection(_connstr, false);
         }
