@@ -42,7 +42,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
                 return Properties.Resources.NotifyIcon1;
             }
         }
@@ -112,7 +112,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
                 return null;
             }
         }
@@ -163,7 +163,7 @@ namespace v2rayN.Handler
         private async Task UpdateTaskRunSubscription(Config config, Action<bool, string> update)
         {
             await Task.Delay(60000);
-            Utils.SaveLog("UpdateTaskRunSubscription");
+            Logging.SaveLog("UpdateTaskRunSubscription");
 
             var updateHandle = new UpdateHandle();
             while (true)
@@ -180,7 +180,7 @@ namespace v2rayN.Handler
                     {
                         update(success, msg);
                         if (success)
-                            Utils.SaveLog("subscription" + msg);
+                            Logging.SaveLog("subscription" + msg);
                     });
                     item.updateTime = updateTime;
                     ConfigHandler.AddSubItem(config, item);
@@ -196,7 +196,7 @@ namespace v2rayN.Handler
             var autoUpdateGeoTime = DateTime.Now;
 
             await Task.Delay(1000 * 120);
-            Utils.SaveLog("UpdateTaskRunGeo");
+            Logging.SaveLog("UpdateTaskRunGeo");
 
             var updateHandle = new UpdateHandle();
             while (true)
