@@ -1,6 +1,5 @@
 ﻿using System.Net;
 using System.Net.Sockets;
-using v2rayN.Base;
 using v2rayN.Mode;
 
 namespace v2rayN.Handler
@@ -29,7 +28,6 @@ namespace v2rayN.Handler
             _updateFunc = update;
 
             Init();
-            Global.StatePort = GetFreePort();
 
             _statisticsV2Ray = new StatisticsV2ray(config, UpdateServerStat);
             _statisticsSingbox = new StatisticsSingbox(config, UpdateServerStat);
@@ -44,7 +42,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
             }
         }
 
@@ -63,7 +61,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
             }
         }
 
@@ -81,6 +79,10 @@ namespace v2rayN.Handler
         {
             GetServerStatItem(_config.indexId);
 
+            if (_serverStatItem is null)
+            {
+                return;
+            }
             if (server.proxyUp != 0 || server.proxyDown != 0)
             {
                 _serverStatItem.todayUp += server.proxyUp;
@@ -88,15 +90,13 @@ namespace v2rayN.Handler
                 _serverStatItem.totalUp += server.proxyUp;
                 _serverStatItem.totalDown += server.proxyDown;
             }
-            if (Global.ShowInTaskbar)
-            {
-                server.indexId = _config.indexId;
-                server.todayUp = _serverStatItem.todayUp;
-                server.todayDown = _serverStatItem.todayDown;
-                server.totalUp = _serverStatItem.totalUp;
-                server.totalDown = _serverStatItem.totalDown;
-                _updateFunc(server);
-            }
+
+            server.indexId = _config.indexId;
+            server.todayUp = _serverStatItem.todayUp;
+            server.todayDown = _serverStatItem.todayDown;
+            server.totalUp = _serverStatItem.totalUp;
+            server.totalDown = _serverStatItem.totalDown;
+            _updateFunc(server);
         }
 
         private void GetServerStatItem(string indexId)
@@ -143,14 +143,12 @@ namespace v2rayN.Handler
                 {
                     return defaultPort;
                 }
-                for (int i = 0; i < 3; i++)
-                {
-                    TcpListener l = new(IPAddress.Loopback, 0);
-                    l.Start();
-                    int port = ((IPEndPoint)l.LocalEndpoint).Port;
-                    l.Stop();
-                    return port;
-                }
+
+                TcpListener l = new(IPAddress.Loopback, 0);
+                l.Start();
+                int port = ((IPEndPoint)l.LocalEndpoint).Port;
+                l.Stop();
+                return port;
             }
             catch
             {
