@@ -6,10 +6,8 @@ using System.Runtime.InteropServices;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Windows;
-using v2rayN.Base;
 using v2rayN.Mode;
 using v2rayN.Resx;
-using v2rayN.Tool;
 
 namespace v2rayN.Handler
 {
@@ -56,7 +54,7 @@ namespace v2rayN.Handler
                             StartInfo = new ProcessStartInfo
                             {
                                 FileName = "v2rayUpgrade.exe",
-                                Arguments = $"\"{fileName}\"",
+                                Arguments = fileName.AppendQuotes(),                                
                                 WorkingDirectory = Utils.StartupPath()
                             }
                         };
@@ -88,7 +86,10 @@ namespace v2rayN.Handler
                     _updateFunc(false, args.Msg);
 
                     url = args.Url;
-                    _ = askToDownload(downloadHandle, url, true);
+                    AskToDownload(downloadHandle, url, true).ContinueWith(task =>
+                    {
+                        _updateFunc(false, url);
+                    });
                 }
                 else
                 {
@@ -141,7 +142,10 @@ namespace v2rayN.Handler
                     _updateFunc(false, args.Msg);
 
                     url = args.Url;
-                    _ = askToDownload(downloadHandle, url, true);
+                    AskToDownload(downloadHandle, url, true).ContinueWith(task =>
+                    {
+                        _updateFunc(false, url);
+                    });
                 }
                 else
                 {
@@ -223,7 +227,7 @@ namespace v2rayN.Handler
                     //more url
                     if (Utils.IsNullOrEmpty(item.convertTarget) && !Utils.IsNullOrEmpty(item.moreUrl.TrimEx()))
                     {
-                        if (!Utils.IsNullOrEmpty(result) && Utils.IsBase64String(result))
+                        if (!Utils.IsNullOrEmpty(result) && Utils.IsBase64String(result!))
                         {
                             result = Utils.Base64Decode(result);
                         }
@@ -247,7 +251,7 @@ namespace v2rayN.Handler
                             }
                             if (!Utils.IsNullOrEmpty(result2))
                             {
-                                if (Utils.IsBase64String(result2))
+                                if (Utils.IsBase64String(result2!))
                                 {
                                     result += Utils.Base64Decode(result2);
                                 }
@@ -274,8 +278,8 @@ namespace v2rayN.Handler
                         int ret = ConfigHandler.AddBatchServers(config, result, id, true);
                         if (ret <= 0)
                         {
-                            Utils.SaveLog("FailedImportSubscription");
-                            Utils.SaveLog(result);
+                            Logging.SaveLog("FailedImportSubscription");
+                            Logging.SaveLog(result);
                         }
                         _updateFunc(false,
                             ret > 0
@@ -327,13 +331,13 @@ namespace v2rayN.Handler
                 }
                 else
                 {
-                    Utils.SaveLog("StatusCode error: " + url);
+                    Logging.SaveLog("StatusCode error: " + url);
                     return;
                 }
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
                 _updateFunc(false, ex.Message);
             }
         }
@@ -350,7 +354,7 @@ namespace v2rayN.Handler
                 foreach (string name in coreInfo.coreExes)
                 {
                     string vName = $"{name}.exe";
-                    vName = Utils.GetBinPath(vName, coreInfo.coreType);
+                    vName = Utils.GetBinPath(vName, coreInfo.coreType.ToString());
                     if (File.Exists(vName))
                     {
                         filePath = vName;
@@ -366,7 +370,7 @@ namespace v2rayN.Handler
                 }
 
                 using Process p = new();
-                p.StartInfo.FileName = filePath;
+                p.StartInfo.FileName = filePath.AppendQuotes();
                 p.StartInfo.Arguments = coreInfo.versionArg;
                 p.StartInfo.WorkingDirectory = Utils.StartupPath();
                 p.StartInfo.UseShellExecute = false;
@@ -399,7 +403,7 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
                 _updateFunc(false, ex.Message);
                 return new SemanticVersion("");
             }
@@ -409,7 +413,7 @@ namespace v2rayN.Handler
         {
             try
             {
-                var gitHubReleases = Utils.FromJson<List<GitHubRelease>>(gitHubReleaseApi);
+                var gitHubReleases = JsonUtils.Deserialize<List<GitHubRelease>>(gitHubReleaseApi);
                 var gitHubRelease = preRelease ? gitHubReleases!.First() : gitHubReleases!.First(r => r.Prerelease == false);
                 var version = new SemanticVersion(gitHubRelease!.TagName);
                 var body = gitHubRelease!.Body;
@@ -524,12 +528,12 @@ namespace v2rayN.Handler
             }
             catch (Exception ex)
             {
-                Utils.SaveLog(ex.Message, ex);
+                Logging.SaveLog(ex.Message, ex);
                 _updateFunc(false, ex.Message);
             }
         }
 
-        private async Task askToDownload(DownloadHandle downloadHandle, string url, bool blAsk)
+        private async Task AskToDownload(DownloadHandle downloadHandle, string url, bool blAsk)
         {
             bool blDownload = false;
             if (blAsk)
@@ -593,7 +597,7 @@ namespace v2rayN.Handler
             {
                 _updateFunc(false, args.GetException().Message);
             };
-            await askToDownload(downloadHandle, url, false);
+            await AskToDownload(downloadHandle, url, false);
         }
 
         private async Task UpdateGeoFile4Singbox(string geoName, Config config, bool needStop, Action<bool, string> update)
@@ -641,7 +645,7 @@ namespace v2rayN.Handler
             {
                 _updateFunc(false, args.GetException().Message);
             };
-            await askToDownload(downloadHandle, url, false);
+            await AskToDownload(downloadHandle, url, false);
         }
 
         #endregion private
