@@ -63,8 +63,8 @@ namespace v2rayN.ViewModels
         {
             if (!Utils.IsNullOrEmpty(normalDNS))
             {
-                var obj = Utils.ParseJson(normalDNS);
-                if (obj != null && obj.ContainsKey("servers") == true)
+                var obj = JsonUtils.ParseJson(normalDNS);
+                if (obj != null && obj["servers"] != null)
                 {
                 }
                 else
@@ -78,7 +78,7 @@ namespace v2rayN.ViewModels
             }
             if (!Utils.IsNullOrEmpty(normalDNS2))
             {
-                var obj2 = Utils.FromJson<Dns4Sbox>(normalDNS2);
+                var obj2 = JsonUtils.Deserialize<Dns4Sbox>(normalDNS2);
                 if (obj2 == null)
                 {
                     UI.Show(ResUI.FillCorrectDNSText);
@@ -87,7 +87,7 @@ namespace v2rayN.ViewModels
             }
             if (!Utils.IsNullOrEmpty(tunDNS2))
             {
-                var obj2 = Utils.FromJson<Dns4Sbox>(tunDNS2);
+                var obj2 = JsonUtils.Deserialize<Dns4Sbox>(tunDNS2);
                 if (obj2 == null)
                 {
                     UI.Show(ResUI.FillCorrectDNSText);
@@ -102,8 +102,8 @@ namespace v2rayN.ViewModels
             ConfigHandler.SaveDNSItems(_config, item);
 
             var item2 = LazyConfig.Instance.GetDNSItem(ECoreType.sing_box);
-            item2.normalDNS = Utils.ToJson(Utils.ParseJson(normalDNS2));
-            item2.tunDNS = Utils.ToJson(Utils.ParseJson(tunDNS2));
+            item2.normalDNS = JsonUtils.Serialize(JsonUtils.ParseJson(normalDNS2));
+            item2.tunDNS = JsonUtils.Serialize(JsonUtils.ParseJson(tunDNS2));
             ConfigHandler.SaveDNSItems(_config, item2);
 
             _noticeHandler?.Enqueue(ResUI.OperationSuccess);
