@@ -79,6 +79,7 @@ namespace v2rayN.Views
                 this.Bind(ViewModel, vm => vm.SelectedSub, v => v.lstGroup.SelectedItem).DisposeWith(disposables);
                 this.Bind(ViewModel, vm => vm.ServerFilter, v => v.txtServerFilter.Text).DisposeWith(disposables);
                 this.BindCommand(ViewModel, vm => vm.AddSubCmd, v => v.btnAddSub).DisposeWith(disposables);
+                this.BindCommand(ViewModel, vm => vm.EditSubCmd, v => v.btnEditSub).DisposeWith(disposables);
 
                 //servers
                 this.BindCommand(ViewModel, vm => vm.AddVmessServerCmd, v => v.menuAddVmessServer).DisposeWith(disposables);
@@ -88,6 +89,7 @@ namespace v2rayN.Views
                 this.BindCommand(ViewModel, vm => vm.AddTrojanServerCmd, v => v.menuAddTrojanServer).DisposeWith(disposables);
                 this.BindCommand(ViewModel, vm => vm.AddHysteria2ServerCmd, v => v.menuAddHysteria2Server).DisposeWith(disposables);
                 this.BindCommand(ViewModel, vm => vm.AddTuicServerCmd, v => v.menuAddTuicServer).DisposeWith(disposables);
+                this.BindCommand(ViewModel, vm => vm.AddWireguardServerCmd, v => v.menuAddWireguardServer).DisposeWith(disposables);
                 this.BindCommand(ViewModel, vm => vm.AddCustomServerCmd, v => v.menuAddCustomServer).DisposeWith(disposables);
                 this.BindCommand(ViewModel, vm => vm.AddServerViaClipboardCmd, v => v.menuAddServerViaClipboard).DisposeWith(disposables);
                 this.BindCommand(ViewModel, vm => vm.AddServerViaScanCmd, v => v.menuAddServerViaScan).DisposeWith(disposables);
@@ -111,7 +113,6 @@ namespace v2rayN.Views
 
                 //servers ping
                 this.BindCommand(ViewModel, vm => vm.MixedTestServerCmd, v => v.menuMixedTestServer).DisposeWith(disposables);
-                this.BindCommand(ViewModel, vm => vm.PingServerCmd, v => v.menuPingServer).DisposeWith(disposables);
                 this.BindCommand(ViewModel, vm => vm.TcpingServerCmd, v => v.menuTcpingServer).DisposeWith(disposables);
                 this.BindCommand(ViewModel, vm => vm.RealPingServerCmd, v => v.menuRealPingServer).DisposeWith(disposables);
                 this.BindCommand(ViewModel, vm => vm.SpeedServerCmd, v => v.menuSpeedServer).DisposeWith(disposables);
@@ -212,8 +213,6 @@ namespace v2rayN.Views
             var IsAdministrator = Utils.IsAdministrator();
             this.Title = $"{Utils.GetVersion()} - {(IsAdministrator ? ResUI.RunAsAdmin : ResUI.NotRunAsAdmin)}";
 
-            spEnableTun.Visibility = IsAdministrator ? Visibility.Visible : Visibility.Collapsed;
-
             //if (_config.uiItem.autoHideStartup)
             //{
             //    WindowState = WindowState.Minimized;
@@ -276,7 +275,7 @@ namespace v2rayN.Views
 
         private void Current_SessionEnding(object sender, SessionEndingCancelEventArgs e)
         {
-            Utils.SaveLog("Current_SessionEnding");
+            Logging.SaveLog("Current_SessionEnding");
             StorageUI();
             ViewModel?.MyAppExit(true);
         }
@@ -330,10 +329,6 @@ namespace v2rayN.Views
                         ViewModel?.AddServerViaClipboard();
                         break;
 
-                    case Key.P:
-                        ViewModel?.ServerSpeedtest(ESpeedActionType.Ping);
-                        break;
-
                     case Key.O:
                         ViewModel?.ServerSpeedtest(ESpeedActionType.Tcping);
                         break;
@@ -343,7 +338,7 @@ namespace v2rayN.Views
                         break;
 
                     case Key.S:
-                        _ = ViewModel?.ScanScreenTaskAsync();
+                        ViewModel?.ScanScreenTaskAsync().ContinueWith(_ => { });
                         break;
 
                     case Key.T:
@@ -479,12 +474,11 @@ namespace v2rayN.Views
             }
 
             var lvColumnItem = _config.uiItem.mainColumnItem.OrderBy(t => t.Index).ToList();
-            for (int i = 0; i < lvColumnItem.Count; i++)
+            var displayIndex = 0;
+            foreach(var item in lvColumnItem)
             {
-                var item = lvColumnItem[i];
-                for (int k = 1; k < lstProfiles.Columns.Count; k++)
-                {
-                    var item2 = (MyDGTextColumn)lstProfiles.Columns[k];
+                foreach (MyDGTextColumn item2 in lstProfiles.Columns) 
+                { 
                     if (item2.ExName == item.Name)
                     {
                         if (item.Width < 0)
@@ -494,7 +488,7 @@ namespace v2rayN.Views
                         else
                         {
                             item2.Width = item.Width;
-                            item2.DisplayIndex = i;
+                            item2.DisplayIndex = displayIndex++;
                         }
                     }
                 }
